2022.2.25

+++ src/views/note/ModifyNote.vue
@@ -36,13 +36,10 @@
             size="mini"
             type="primary"
             icon="el-icon-search"
+            @click="checkTheNote(scoped.row.id)"
           ></el-button>
           <!-- 修改笔记信息按钮 -->
-          <el-button
-            size="mini"
-            type="info"
-            icon="el-icon-edit"
-          ></el-button>
+          <el-button size="mini" type="info" icon="el-icon-edit" @click="modifyTheNote(scoped.row.id)"></el-button>
           <!-- 删除笔记按钮 -->
           <el-button
             size="mini"
@@ -62,11 +59,54 @@
       :total="allTotal"
     >
     </el-pagination>
+    <!-- 查看对话框 -->
+    <el-dialog title="笔记内容" :visible.sync="noteDiaVisible" width="60%">
+      <span v-html="oneNote.note_content"></span>
+      <span slot="footer" class="dialog-footer">
+        <el-button @click="noteDiaVisible = false">取 消</el-button>
+        <el-button type="primary" @click="noteDiaVisible = false"
+          >确 定</el-button
+        >
+      </span>
+    </el-dialog>
+    <!-- 修改对话框 -->
+    <el-dialog title="笔记内容" :visible.sync="modifyNoteDiaVisible" width="60%">
+      <el-form
+        ref="ontnoteform"
+        :rules="ontNoteRules"
+        :model="oneNote"
+        label-width="80px"
+      >
+        <el-form-item label="标题" prop="note_title">
+          <el-input v-model="oneNote.note_title"></el-input>
+        </el-form-item>
+        <el-form-item label="分类" prop="note_sort">
+          <el-input v-model="oneNote.note_sort"></el-input>
+        </el-form-item>
+        <el-form-item label="描述" prop="note_describe">
+          <el-input v-model="oneNote.note_describe"></el-input>
+        </el-form-item>
+        <el-form-item label="内容" prop="note_content">
+          <el-input
+            type="textarea"
+            rows="10"
+            resize="none"
+            v-model="oneNote.note_content"
+          ></el-input>
+        </el-form-item>
+      </el-form>
+      <span slot="footer" class="dialog-footer">
+        <el-button @click="modifyNoteDiaVisible = false">取 消</el-button>
+        <el-button type="primary" @click="modifyTheNotes()"
+          >修 改</el-button
+        >
+      </span>
+    </el-dialog>
   </div>
 </template>
 
 <script>
-import { getAllNote, deleteNote, getSomeNote } from "@/api/index";
+import { getAllNote, deleteNote, getSomeNote, getOneNote,modifyTheNote } from "@/api/index";
 export default {
   created() {
     this.getSomeNote(this.pageSizes, this.currentPage);
@@ -81,6 +121,39 @@ export default {
       currentPage: 1,
       // 当前页显示数
       pageSizes: 10,
+      // 查看对话框的显示与隐藏
+      noteDiaVisible: false,
+      // 修改对话框的显示与隐藏
+      modifyNoteDiaVisible: false,
+      // 单个笔记信息
+      oneNote: {},
+      // 修改验证规则
+      ontNoteRules: {
+        note_title: [
+          { required: true, message: "请输入笔记标题", trigger: "blur" },
+          {
+            min: 3,
+            max: 100,
+            message: "长度在 3 到 100 个字符",
+            trigger: "blur",
+          },
+        ],
+        note_sort: [
+          { required: true, message: "请输入笔记分类", trigger: "blur" },
+        ],
+        note_describe: [
+          { required: true, message: "请输入笔记描述", trigger: "blur" },
+          {
+            min: 3,
+            max: 100,
+            message: "长度在 3 到 100 个字符",
+            trigger: "blur",
+          },
+        ],
+        note_content: [
+          { required: true, message: "请输入笔记描述", trigger: "blur" },
+        ]
+      },
     };
   },
   methods: {
@@ -97,18 +170,56 @@ export default {
     },
     // 删除笔记按钮
     async deleteNote(id) {
-      // 如何获取要删除的那一行的id(tableNoteData.id)：通过作用域插槽的slot-scope="scoped"，scoped是一个对象，存储了这个子组件的所有属性，再通过scoped.row.id获取id
-      const result = await deleteNote(id);
-      if (result.meta.status == 200) {
-        this.$message.success("删除成功！");
-        this.getSomeNote(this.pageSizes, this.currentPage);
-      }
+      this.$confirm("此操作将永久删除该笔记, 是否继续?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning",
+      })
+        .then(() => {
+          // 如何获取要删除的那一行的id(tableNoteData.id)：通过作用域插槽的slot-scope="scoped"，scoped是一个对象，存储了这个子组件的所有属性，再通过scoped.row.id获取id
+          deleteNote(id)
+            .then((resolve) => {
+              this.$message.success("删除成功！");
+              this.getSomeNote(this.pageSizes, this.currentPage);
+            })
+            .catch((error) => {
+              this.$message.success("删除失败！");
+            });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除",
+          });
+        });
     },
     // 当前页数改表时触发
     pageChange(newPage) {
-      this.currentPage = newPage
+      this.currentPage = newPage;
       this.getSomeNote(10, newPage);
     },
+    // 查看笔记按钮
+    async checkTheNote(id) {
+      const data = await getOneNote(id);
+      this.oneNote = data;
+      this.noteDiaVisible = true;
+    },
+    // 修改笔记按钮
+    async modifyTheNote(id) {
+      this.modifyNoteDiaVisible = true
+      this.oneNote = await getOneNote(id);
+    },
+    // 对话框里的修改按钮
+    modifyTheNotes() {
+      console.log(this.oneNote);
+      modifyTheNote(this.oneNote.id,this.oneNote.note_title,this.oneNote.note_describe, this.oneNote.note_content, this.oneNote.note_sort).then(resolve => {
+        this.$message.success("修改成功！")
+        this.getSomeNote(this.pageSizes, this.currentPage);
+        this.modifyNoteDiaVisible = false
+      }).catch(error => {
+        this.$message.error("修改失败！")
+      })
+    }
   },
 };
 </script>
@@ -123,7 +234,7 @@ export default {
   .el-pagination {
     position: absolute;
     left: 50%;
-    transform: translateX(-50%);
+    transform: translateX(-60%);
   }
 }
 </style>
